// webapp/app/containers/Organizations/types.ts
export interface IDavinciUser {
  id: number
  username: string
  email: string
  role: number
  avatar: string
}
export interface IOrganization {
  id: number
  name: string
  role?: number
  avatar: string
  roleNum?: number
  memberNum?: number
  projectNum?: number
  description: string
  memberPermission?: number
  allowCreateProject?: number
  allowChangeVisibility?: number
  allowDeleteOrTransferProject?: number
}

export interface IOrganizationMember {
  id: number
  user: IDavinciUser
}

export interface IOrganizationRole {
  id: number
  name: string
  description: string
}
// ...
export interface IMembers {
  id: number
  user: IDavinciUser
  roles?: object[]
}